Change prices in ExchangeItem to be doubles

+++ src/main/kotlin/dev/the/mag/exchangebrokerbackend/dto/ExchangeItemDto.kt
@@ -7,9 +7,9 @@ data class ExchangeItemDto (
 
     var name: String,
     var desc: String,
-    var price: Int,
+    var price: Double,
 
-    var soldFor: Int?,
+    var soldFor: Double?,
     var pending: Boolean?,
 
     var earningPercent: Int,
